Creating testing project.  Modifed namespaces for better compliance.  Updated DupeFinder to be in it's own class. Added events so application interface can be dealt with separately.
