+++ DupeFinder/dupeFinder.csproj
