+++ DupeFinderTest/ProveTestWorksTest.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace DupeFinderTest
+{
+    public class ProveTestWorksTest
+    {
+        [Fact]
+        public void ProveTestWorks_TestPasses()
+        {
+            Assert.True(true);
+        }
+    }
+}
