+++ dupeFinder.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dupeFinder", "dupeFinder.csproj", "{0C1BB45F-5B88-447A-8F4F-C7BE6A003616}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DupeFinderTest", "DupeFinderTest\DupeFinderTest.csproj", "{BBD9D5B3-1A27-4343-BB67-9004E9814627}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dupeFinder", "DupeFinder\dupeFinder.csproj", "{6981E94B-CAE8-4A82-A4A0-4C9EE54ABE53}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0C1BB45F-5B88-447A-8F4F-C7BE6A003616}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0C1BB45F-5B88-447A-8F4F-C7BE6A003616}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C1BB45F-5B88-447A-8F4F-C7BE6A003616}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0C1BB45F-5B88-447A-8F4F-C7BE6A003616}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BBD9D5B3-1A27-4343-BB67-9004E9814627}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BBD9D5B3-1A27-4343-BB67-9004E9814627}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBD9D5B3-1A27-4343-BB67-9004E9814627}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BBD9D5B3-1A27-4343-BB67-9004E9814627}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6981E94B-CAE8-4A82-A4A0-4C9EE54ABE53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6981E94B-CAE8-4A82-A4A0-4C9EE54ABE53}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6981E94B-CAE8-4A82-A4A0-4C9EE54ABE53}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6981E94B-CAE8-4A82-A4A0-4C9EE54ABE53}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
